requirements.txt
fastapi
uvicorn
ffmpeg-python
python-multipart
python-dotenv
pika
